# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MediVault API
  version: 0.21.0
  description: |
    Patient-centric medical records platform. Patients own their data, doctors
    get scoped time-bound access, caregivers act on behalf of linked patients.

    **Authentication.** All endpoints except `/auth/*` and `/emergency/:patientCode`
    require a Bearer JWT in the `Authorization` header. Tokens are issued at
    `/auth/login` and `/auth/verify-otp`; refresh tokens rotate at `/auth/refresh`.

    **Roles.** `PATIENT`, `DOCTOR`, `CAREGIVER`, `ADMIN`. Endpoints declare which
    roles can call them via the role-gate preHandler. Mismatched role returns
    `403 forbidden`.

    **Error shape.** All errors return `{ "error": "<code>", "message": "<human>"}`.
    Validation errors additionally include `issues: [{path, message, code}]`.

  contact:
    name: MediVault
    url: https://medivault.app
servers:
  - url: http://localhost:3001/v1
    description: Local development
  - url: https://api.medivault.app/v1
    description: Production (not deployed yet)
...
paths:
...
  /access/request:
    post:
      tags: [Access]
      summary: 'Doctor: request access to a patient by patient code'
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientCode]
              properties:
                patientCode: { type: string, pattern: '^MVK-\d{4}-\d{5,}$' }
                note: { type: string }
      responses:
        '201': { description: Request created }
        '403': { description: Doctor not verified, $ref: '#/components/responses/Error' }
        '404': { description: Patient not found, $ref: '#/components/responses/Error' }
        '409': { description: Duplicate active request, $ref: '#/components/responses/Error' }
  /access/incoming:
    get:
      tags: [Access]
      summary: 'Patient: list incoming access requests + grants'
      security: [{ bearer: [] }]
      responses: { '200': { description: List of access permissions } }
  /access/outgoing:
    get:
      tags: [Access]
      summary: 'Doctor: list own outgoing requests + active grants'
      security: [{ bearer: [] }]
      responses: { '200': { description: List of access permissions } }
  /access/{id}/approve:
    post:
      tags: [Access]
      summary: 'Patient: approve a pending request'
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [duration]
              properties: { duration: { type: string, enum: [HOURS_24, DAYS_7, DAYS_30, PERMANENT] } }
      responses: { '200': { description: Approved } }
  /access/{id}/deny:
    post:
      tags: [Access]
      summary: 'Patient: deny a pending request'
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Denied } }
  /access/{id}/revoke:
    post:
      tags: [Access]
      summary: 'Patient: revoke an active grant'
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Revoked } }
  /patients/{patientCode}:
    get:
      tags: [Patients (Doctor view)]
      summary: 'Doctor: get a patient profile (requires active access)'
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: patientCode, required: true, schema: { type: string } } ]
      responses:
        '200': { description: Patient profile }
        '403': { description: No active access, $ref: '#/components/responses/Error' }
  /patients/{patientCode}/records:
    get:
      tags: [Patients (Doctor view)]
      summary: 'Doctor: list a patient''s records'
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: patientCode, required: true, schema: { type: string } } ]
      responses: { '200': { description: Records list } }
  /patients/{patientCode}/medications:
    get:
      tags: [Patients (Doctor view)]
      summary: 'Doctor: list a patient''s medications'
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: patientCode, required: true, schema: { type: string } } ]
      responses: { '200': { description: Medications list } }
  /patients/{patientCode}/drug-interactions:
    get:
      tags: [Patients (Doctor view)]
      summary: 'Doctor: cached drug-interaction check (does not charge doctor''s AI rate limit)'
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: patientCode, required: true, schema: { type: string } } ]
      responses: { '200': { description: Interactions } }
  /patients/{patientCode}/summary:
    get:
      tags: [Patients (Doctor view)]
      summary: 'Doctor: get cached AI summary; generate if missing'
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: patientCode, required: true, schema: { type: string } } ]
      responses: { '200': { description: AI summary } }
...
  /patients/{patientCode}/ask:
    post:
      tags: [Patients (Doctor view)]
      summary: 'Doctor: ask a free-form question about the patient'
      security: [{ bearer: [] }]
      parameters: [ { in: path, name: patientCode, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties: { question: { type: string, minLength: 3, maxLength: 1000 } }
      responses:
        '200':
          description: AI answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer: { type: string }
                  modelId: { type: string }
                  generatedAt: { type: string, format: date-time }
                  tokens: { type: object, properties: { input: { type: integer }, output: { type: integer } } }
                  disclaimer: { type: string }
        '429': { description: AI rate limit reached, $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string, description: 'machine-readable code, e.g. forbidden, validation_error' }
        message: { type: string }
        issues:
          type: array
          description: 'Present on validation errors only'
          items:
            type: object
            properties:
              path: { type: array, items: { type: string } }
              message: { type: string }
              code: { type: string }
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }